requirements-base.txt
mq-transport[sqs]@git+https://github.com/Bondifuzz/mq-transport.git@0.1.0
aiohttp==3.8.1
aioarangodb==0.1.2
PyYAML==5.4.1
coloredlogs==15.0.1